Fix sign conversions in the network layer and dialogs
Connection timing is signed in both games: Set_Timing, the retry delay,
retry limit and timeout are int32_t with -1 as the documented no-limit
value, response times are int32_t, and the connection clock and queue
timestamps are int64_t, so the tick count survives long uptimes.

CRC inputs are converted to the CRC's unsigned type, keeping the bits that
detect desync; TD's CRC and event layout stay as they are, since events
travel by sizeof(EventClass). Packet and serial structs keep their widths.
The null-modem, IPX and Westwood Online dialogs use signed positions and
timers, convert at SafeCopy, strncpy and COM interface calls, and
Net_Reconnect_Dialog takes an int, matching its %02d format.

+++ src/ra/combuf.h
@@ -53,6 +53,8 @@
 #ifndef CNC_RED_ALERT_RA_COMBUF_H_
 #define CNC_RED_ALERT_RA_COMBUF_H_
 
+#include <cstdint>
+
 /*
 ********************************** Defines **********************************
 */
@@ -62,9 +64,9 @@ This is one output queue entry
 typedef struct {
   unsigned int IsActive : 1;  // 1 = this entry is ready to be processed
   unsigned int IsACK : 1;     // 1 = ACK received for this packet
-  long FirstTime;    // time this packet was first sent
-  long LastTime;     // time this packet was last sent
-  long SendCount;    // # of times this packet has been sent
+  int64_t FirstTime;          // time this packet was first sent
+  int64_t LastTime;           // time this packet was last sent
+  int SendCount;              // # of times this packet has been sent
   int BufLen;                 // size of the packet stored in this entry
   char* Buffer;               // the data packet
   int ExtraLen;               // size of extra data
@@ -131,9 +133,9 @@ class CommBufferClass {
   /*
   ....................... Response time routines ........................
   */
-  void Add_Delay(unsigned long delay);  // accumulates response time
-  unsigned long Avg_Response_Time();    // gets mean response time
-  unsigned long Max_Response_Time();    // gets max response time
+  void Add_Delay(int64_t delay);  // accumulates response time
+  int32_t Avg_Response_Time();    // gets mean response time
+  int32_t Max_Response_Time();    // gets max response time
   void Reset_Response_Time();           // resets computations
 
   /*
@@ -159,10 +161,10 @@ class CommBufferClass {
   /*
   ....................... Response time variables .......................
   */
-  unsigned long DelaySum = 0;  // sum of last 4 delay times
-  unsigned long NumDelay = 0;  // current # delay times summed
-  unsigned long MeanDelay = 0;  // current average delay time
-  unsigned long MaxDelay = 0;  // max delay ever for this queue
+  int64_t DelaySum = 0;   // sum of last 4 delay times
+  int NumDelay = 0;       // current # delay times summed
+  int64_t MeanDelay = 0;  // current average delay time
+  int64_t MaxDelay = 0;   // max delay ever for this queue
 
   /*
   ........................ Send Queue variables .........................
