+++ src/ra/seqconn.cc
@@ -78,9 +78,8 @@
  *=========================================================================*/
 SequencedConnClass::SequencedConnClass(int numsend, int numreceive, int maxlen,
                                        unsigned short magicnum,
-                                       unsigned long retry_delta,
-                                       unsigned long max_retries,
-                                       unsigned long timeout)
+                                       int32_t retry_delta,
+                                       int32_t max_retries, int32_t timeout)
     : ConnectionClass(maxlen, magicnum, retry_delta, max_retries, timeout) {
   NumRecNoAck = 0;
   NumRecAck = 0;
